feat: Add incatationInProgess flag to Trantorian class
The code changes add a new boolean flag `incatationInProgess` to the `Trantorian` class in the `Trantorian.hpp` file. This flag is used to track whether an incantation is in progress for a particular `Trantorian` object.

Based on the recent user commits and repository commits, it seems that the changes are related to implementing the GUI for the Zappy game. The commits include merging branches, refactoring code, and updating the display of teams, chat, fps, and time unit.

Considering the established commit message conventions in the repository, the suggested commit message is "feat: Add incatationInProgess flag to Trantorian class".